// lib/generated/prisma/models/AlasanLembur.ts
export type AlasanLemburModel = runtime.Types.Result.DefaultSelection<Prisma.$AlasanLemburPayload>
// ...
export type $AlasanLemburPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "AlasanLembur"
  objects: {}
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: number
    description: string
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["alasanLembur"]>
  composites: {}
}